Create pension.py
The pension fund for a small company provides for a man to receive 5000 TL a month if he is over 65 and an extra 2000 TL if he is over 70. A woman receives 4500 TL a month if she is over 60, with an extra 2500 TL if she is over 65. Write a function  that will read in the gender and age of a person and write out the amount of their pension. If the person is under pensionable age a suitable message should be written.